+++ scripts/run-cool2multirescool.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+input=$1
+ncores=$2
+cooler coarsegrain $input -n $ncores
+