+++ VERSION_HISTORY.md
@@ -1,3 +1,9 @@
+### v4
+* updated cooler (Issues with pairix file input have been resolved)
+* run-cool2multirescool.sh available now.
+* pairsqc is installed now. R and Java are also installed as dependency (both are needed for nozzle).
+* Usage for run-cooler.sh changed (ncores added)
+
 ### v3
 * updated cooler